+++ CHANGES.md
@@ -1,2 +1,3 @@
 #### [unreleased]
-* initial pass
+* add/remove Collections of Git Updater Update API Servers
+* add tab in Plugin Install page for plugins in collections